backend/requirements.txt
fastapi
uvicorn
pydantic
PyYAML
pytest
httpx
PyJWT
bcrypt
